// web/web-client/src/types/media.ts
// tslint:disable
/**
 * mediaTrack
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface Media
 */
export interface Media {
    /**
     * 
     * @type {number}
     * @memberof Media
     */
    id: number;
    /**
     * name
     * @type {string}
     * @memberof Media
     */
    name: string;
    /**
     * 
     * @type {string}
     * @memberof Media
     */
    path: string;
    /**
     * 
     * @type {string}
     * @memberof Media
     */
    descript?: string;
    /**
     * 创建时间
     * @type {string}
     * @memberof Media
     */
    createdAt?: string;
    /**
     * 更新时间
     * @type {string}
     * @memberof Media
     */
    updatedAt?: string;
    /**
     * 删除时间
     * @type {string}
     * @memberof Media
     */
    deletedAt?: string;
    /**
     * 类型
     * @type {number}
     * @memberof Media
     * 1: local, 2: url
     */
    type: number;
}